breaking up the js code
