+++ doc/algorithm.md
@@ -18,7 +18,7 @@ The calculations and algorithm are coming from the following websites:
 Basically, you get a word (English), then scramble the characters, then hopefully you find 
 another correct (English) word. 
 
-First, you might try to list out everything possible combination. 
+First, you might try to list out everything possible combination. Let's try 3 letters.
 
 ```
 cat
@@ -29,7 +29,7 @@ tca
 tac
 ```
 
-So a 3 letters long word can have 6 different combinations.
+So a 3 letters long word can have 6 different combinations. What about 4 letters.
 
 ```
 aces
@@ -58,13 +58,14 @@ ecsa
 cesa
 ```
 
-And 4 letters long word can have up to 24 different combinations. One extra letter create quite a big jump in numbers of possibilities. 
+A 4 letters long word can have up to 24 different combinations. One extra letter create quite a big jump in numbers of possibilities. 
 
 The formula to calculation all possible combination is like this 
 
 ```
-3 + 2 + 1 = 6
-4 + 3 + 2 + 1 = 6
+3 * 2 = 6
+4 * 3 * 2 = 24
+5 * 4 * 3 * 2 = 120
 ``` 
 
 For proper mathematics equation representation can see the reference section above. 
@@ -74,18 +75,28 @@ Now we need to use computer language to calculate the result. Here I am using Ja
 ```js 
 // version 1
 // calculate possible combination total 
-function cpct(n, total = 0) {
-    // exit cause 
-    if (n === 0) {
-        return total
+function getCombinationTotal(n, total = 0) {
+    // exit cause
+    if (n === 1) {
+       return total
     }
-    // init 
+    // init
     if (total === 0) {
-        total = n 
+       total = n
     }
-    total = n + (n-1)
-    n-- 
-    return cpct(n, total)
+    n--
+    total *= n
+    return getCombinationTotal(n, total)
 }
 ```
+Now here is another problem, what if we put a 15 letters long word, 
+how many combination would that be?
 
+**1,307,674,368,000**
+
+Thats over a trillion possibilities ...
+
+In the version 1 of this program, I only use a random generator to half solve this problem, 
+basically it's like rolling a dice, and see if I bet on the right one. 
+
+To be continue ...
