+++ doc/changelog.md
@@ -1,0 +1,10 @@
+# One possible solution to the problem we have
+
+If you have run the program several times, you might encounter the famous `stackoverflow` error. 
+Basically, its too much calculation to do, your machine ran out of memory.
+
+This is why we only limit the maximum letters long to 7. 
+Even that it could have to calculate at least 5,040 times;
+It could be more because the random generator might came up with the same combination.
+
+How do we solve this problem?
