+++ js/math.js
@@ -6,9 +6,9 @@
  * @param {int} total to hold the total number
  * @return {int} the total possible combination
  */
-function cpct(n, total = 0) {
+function getCombinationTotal(n, total = 0) {
     // exit cause
-    if (n === 0) {
+    if (n === 1) {
        return total
     }
     // init
@@ -16,25 +16,19 @@ function cpct(n, total = 0) {
        total = n
     }
     n--
-    total += n
-    return cpct(n, total)
+    total *= n
+    return getCombinationTotal(n, total)
 }
 
-function cpct_0(n, total = 0) {
-    // exit cause
-    if (n === 0) {
-        return total
-    }
-    // init
-    if (total === 0) {
-        total = n
-    }
-    total = n + (n-1)
-    n--
-    return cpct(n, total)
+module.exports = {
+    getCombinationTotal
 }
 
-const result1 = cpct_0(4)
-const result2 = cpct_0(5)
+/*
+let result5 = getCombinationTotal(5)
+let result6 = getCombinationTotal(6)
+let result7 = getCombinationTotal(7)
 
-console.log(result1, result2)
+
+console.log(result5, result6, result7)
+*/
